Add information on licence and OSM usage permission
This commit adds information on the licence and the nature of the
permission to use the layer in OSM to the configurations for the global
aerial imagery mosaics.

As there is no fitting SPDX licence id for non open licences I've used
"COMMERCIAL" which seems to be fit well and is reasonably diplomatic.