Add support for HTTP header indicating invalid tiles (#552)
This adds fields that allow specifying a HTTP header and associated value that indicates an invalid tile, this is mainly useful for layers that return error tiles instead of 404s. This mirrors a similar facility of JOSM.

Note, this could naturally be modelled as a JSON object or an array instead of two string fields, no particular preference on my behalf.