Juno GPS tiles for NYC (#625)
This adds a new overlay layer that looks exactly like the OpenStreetMap's GPS tile layer, except that it covers only New York City, has much more tracks (30-40 thousand) and only tracks from a single day, always at most two days old. It is updated daily. The license is CC-BY-SA 4.0 with an explicit permission to use for OSM.

See https://gps-tile.junolab.net/ for an overview and terms of use.

In a week or two we open in New Jersey, so it will have more data. Not sure if putting it in the `ny` directory is correct, but it certainly doesn't cover the whole of US.

I have also updated the `requirements.txt` file, since version 0.13.3 of the `transifex-client` has been pulled from pypi.