Create Kanton_Solothurn_SOGIS_DTM_2018.geojson (#810)
* Create Kanton_Solothurn_SOGIS_DTM_2018.geojson

I saw, that the 2019 relief only provides the southern half of the canton. The 2018 relief is the north, including most of Basel Land as well. This would also require modifications to the polygon .. but I do not know how to do that.

* Update Kanton_Solothurn_SOGIS_DTM_2018.geojson

Changed the polygon to reflect the actual map better