Update MD boundary
The previous boundary was not very accurate and it did not even cover the full territory of the state of Maryland. This makes it impossible to set the "best" as true.

Now the boundary near Potomac River (MD|VA), Mason-Dixon line (MD|PA), MD|DC as well as MD|DE have been carefully checked. MD|WV side has not been revised very well because WV does not have its own orthoimagery available for OSM, so it does not hurt to give others a slight chance to use it.