added ACC 2018 ortho imagery (#1158)
* added ACC 2018 ortho imagery

* added pdf and deleted duplicated .txt file

I added the license permission for using the ACC 2018 ortho imagery as a .pdf file. I also deleted the text file which I realized was the geojson file I created at first.

* fixed 't's in .geojson

changed uppercase t's to lowercase t's for 'true' in the geojson file. Also added in an quotation mark.

* Update ACC_Data.geojson

Made a better geometry for bounding the county in.

* Fixed geoJSON issues

changed min zoom level to 8 and fixed permalink to license

* Update ACC_data.geojson

* Add files via upload

* Delete Athens-Clarke_County_Data.pdf

* Update ACC_Data.geoJSON

fixed link to right file

* update license url

Co-authored-by: Andrew Harvey <andrew@alantgeo.com.au>