Create CzechCUZK_ortophoto.geojson
Some very useful layers from CUZK (State Administration of Land Surveying and Cadastre) have become legally available for OSM mapping. The OSM CZ have managed to get an explicit approval on top of a CC-BY-4.0 license for these layers. This pull request is to add an excellent orthophoto of the whole Czech Republic. I believe that it is the best option for the region in terms of alignment accuracy, resolution and update frequency, so I have marked it accordingly. The earlier added IPR aerial imagery is even better but that one is limited to the area of Prague only.
The second layer to be added is a digital model of the surface of the terrain. The description is inside the file.