Fixed start/end dates don't make really make sense
All three products are continously updated, the terrain ones in a six
year cycle the orthofoto faster. Making the start and end dates quite
meaningless outside of a rough indication from when the data is.

Maybe it would be a better idea to allow specification of an
"update_cycle" value, which would allow the user at least to determine
the maximum age of the imagery they are using.